# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Library Management API
  description: REST APIs for managing books, members, and borrow/return operations
  version: 1.0.0

servers:
  - url: http://localhost:8081

tags:
  - name: Books
  - name: Members
  - name: Borrowing

paths:

  /api/books:
    post:
      tags: [Books]
      operationId: createBook
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '201':
          description: Book created

    get:
      tags: [Books]
      operationId: listBooks
      summary: List all books
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookResponse'

  /api/books/{id}:
    put:
      tags: [Books]
      operationId: updateBook
      summary: Update book details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '200':
          description: Book updated

  /api/members:
    post:
      tags: [Members]
      operationId: createMember
      summary: Create a member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberRequest'
      responses:
        '201':
          description: Member created

    get:
      tags: [Members]
      operationId: getMembersList
      summary: List all members
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberResponse'

  /api/members/{memberId}/borrowed-books:
    get:
      tags: [Borrowing]
      operationId: listBorrowedBooks
      summary: List borrowed books for a member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Borrowed books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BorrowedBookResponse'

  /api/borrow:
    post:
      tags: [Borrowing]
      operationId: borrowBook
      summary: Borrow a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowRequest'
      responses:
        '200':
          description: Book borrowed successfully

  /api/return/{transactionId}:
    post:
      tags: [Borrowing]
      operationId: returnBook
      summary: Return a borrowed book
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Book returned successfully

components:
  schemas:

    BookRequest:
      type: object
      required: [title, totalCopies]
      properties:
        title:
          type: string
          example: Clean Code
        author:
          type: string
          example: Robert C. Martin
        isbn:
          type: string
          example: 9780132350884
        totalCopies:
          type: integer
          example: 5

    BookResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        totalCopies:
          type: integer
        availableCopies:
          type: integer

    MemberRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john@example.com
        phone:
          type: string
          example: 9876543210

    MemberResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        phone:
          type: string

    BorrowRequest:
      type: object
      required: [bookId, memberId]
      properties:
        bookId:
          type: integer
          format: int64
          example: 1
        memberId:
          type: integer
          format: int64
          example: 10

    BorrowedBookResponse:
      type: object
      properties:
        bookId:
          type: integer
          format: int64
        bookTitle:
          type: string
          example: Clean Code
        borrowedAt:
          type: string
          format: date-time